+++ CHANGELOG.md
@@ -1,4 +1,15 @@
-﻿## 1.0.1
+﻿## Changelog
+
+## 1.1.0
+- Added team battles: companions can now be selected to join a practice fight as allies
+- Companions will also use horses when appropriate, same as the player
+
+## 1.0.2
+- If cavalry units are selected they will equip their horse and the player will too
+- Added JSON configuration file logic to allow users to customize settings
+- Added MaxTrainingTroops setting to config file
+
+## 1.0.1
 
 - Added troop wounding system: defeated training opponents become wounded and retain status through promotion.
 - Added wounded troop restriction: wounded units cannot be selected for training fights.